requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
# PyPDF2
pdfplumber
faiss-cpu
langchain_google_genai
langchain-community
rank_bm25